0x0D-preprocessor: reverted code

+++ 0x0D-preprocessor/3-function_like_macro.h
@@ -1,5 +1,5 @@
 #ifndef MAIN
 #define MAIN
-#define ABS(X) ((X) < 0 ? -((unsigned int)X) : (X))
+#define ABS(X) ((X) < 0 ? -(X) : (X))
 typedef int to_make_iso_happy;
 #endif
